chore(sdk): Move reference definition to typedefs
