+++ packages/grafbase-sdk/src/grafbase-schema.ts
@@ -2,7 +2,7 @@ import { Model, ModelFields } from './model'
 import { RelationDefinition, RelationRef } from './relation'
 import { Enum, EnumShape } from './enum'
 import { Type, TypeFields } from './type'
-import { ReferenceDefinition } from './reference'
+import { ReferenceDefinition } from './typedefs/reference'
 import { Union } from './union'
 import { Interface, InterfaceFields } from './interface'
 import { Query, QueryInput } from './query'
