+++ packages/grafbase-sdk/src/typedefs/reference.ts
@@ -1,9 +1,9 @@
-import { AuthRuleF } from './auth'
-import { Enum } from './enum'
-import { ListDefinition } from './field/list'
-import { Type } from './type'
-import { AuthDefinition } from './typedefs/auth'
-import { ResolverDefinition } from './typedefs/resolver'
+import { AuthRuleF } from '../auth'
+import { Enum } from '../enum'
+import { ListDefinition } from '../field/list'
+import { Type } from '../type'
+import { AuthDefinition } from './auth'
+import { ResolverDefinition } from './resolver'
 
 export class ReferenceDefinition {
   referencedType: string
