+++ packages/grafbase-sdk/src/typedefs/resolver.ts
@@ -1,8 +1,8 @@
 import { AuthRuleF } from '../auth'
-import { ReferenceDefinition } from '../reference'
 import { AuthDefinition } from './auth'
 import { CacheDefinition, CacheParams, TypeLevelCache } from './cache'
 import { DefaultDefinition } from './default'
+import { ReferenceDefinition } from './reference'
 import { ScalarDefinition } from './scalar'
 import { SearchDefinition } from './search'
 import { UniqueDefinition } from './unique'
